--- scripts/extract_msg
@@ -1,5 +1,0 @@
-#!/usr/bin/env python
-from extract_msg.__main__ import main
-
-if __name__ == '__main__':
-    main()
